Server side type changes for SynthDef2
Changes sizes of struct members to match SynthDef2
and make necessary changes in other code. Also 
some cleanup. Debug code still in place.

+++ include/plugin_interface/SC_Unit.h
@@ -39,7 +39,7 @@ struct Unit
 	struct World *mWorld;
 	struct UnitDef *mUnitDef;
 	struct Graph *mParent;
-	uint16 mNumInputs, mNumOutputs;
+	uint32 mNumInputs, mNumOutputs; // changed from uint16 for synthdef ver 2
 	int16 mCalcRate;
 	int16 mSpecialIndex;		// used by unary and binary ops
 	int16 mParentIndex;
